refactor: move db accessing attempt fns to handler methods
